updated the doc with the correct CLI default exit time
The default CLI exit time is 10 sec as per the configuration. I followed the whole readme and found that this is a big issue and very hard to notice and can be a big impediment while following this doc. 
If you want to update this information by yourself, you can do it in your language. I've added the important details already as a edit to this document.